give each ambiguous ?: a single operand type
GCC rejects a conditional whose two branches have different types when each converts
to the other: there is no reason to prefer one conversion over the other, so there is
no composite type. MSVC picks one and says nothing, which is how ten of these
accumulated. Every type involved has both a non-explicit converting constructor and a
conversion operator back:

  CPtr / CObj / CMObj   CPtrBase( T * )       operator T *()
  CDynamicCast<T>       CDynamicCast( TT * )  operator T *()
  CDBPtr<T>             CDBPtr( const T * )   operator const T *()
  SAIAngle              SAIAngle( int )       operator int()
  CVariant              CVariant( int ), ...  operator int(), ...

Which conversion MSVC was choosing matters, because four of these are in AILogic and
Common_RTS_AI, where the simulation has to stay in lockstep. Compiling the three
shapes against cl.exe and reading back the decltype says MSVC picks the class type in
all of them, so each fix below is written to produce what the class-type composite
already produced.

For SAIAngle the class-type composite built an SAIAngle from the int operand, which
truncates to the uint16_t the struct holds; the result then converted straight back to
an integer at the assignment. Naming that uint16_t on the other operand gives an
integer composite that truncates at the same assignment instead, for the same value.

For CVariant in libdb/TypeDef.cpp the composite was CVariant, so the empty-list branch
stayed VT_NULL. Making the other operand a CVariant keeps that. Had MSVC picked int,
the empty branch would have gone through operator int(), which maps VT_NULL to 0, and
written VT_INT 0 instead - so this one was worth checking rather than assuming.

For the smart pointers, GetPtr() is the same function the implicit conversion calls, so
the value is unchanged. What does change is that the class-type composite used to build
a temporary wrapper around the *raw* operand and destroy it at the end of the
expression. In Parser/LangNodesDefinitions.h that temporary was a CPtr around `this`,
and CPtrBase's release calls ReleaseRef, which deletes the object when both counters
reach zero. CNodesList holds its nodes by CPtr, so a node reached through a list always
has a live reference and the round trip never actually freed anything, but it was one
reference away from returning a pointer to a deleted object. The same round trip was in
LangNodesDefinitions.cpp. In Stats_B2_M1 the temporary was a CDBPtr around what
NDb::Get had just returned, whose conversion back re-ran LoadObject on an object that
was already loaded.

Nine sites came from GCC, which attempted 1276 translation units in a keep-going build.
The tenth, ED_B2_M1/XMLExport.cpp, is in the editor, which GCC never compiles.

This class cannot be swept the way the typename ones were. The branch types nearly
always come from another file, and `wrapper ? x : 0` is well-formed because 0 is a null
pointer constant, so a source-level scan is almost entirely false positives - it
produced thirty candidates here, of which one was real. Expect more of these to surface
as more of the tree reaches the compiler.

+++ Versions/Temporary/Engine/Sources/AILogic/CommonStates.cpp
@@ -1552,7 +1552,7 @@ void CCommonSwarmState::Segment()
 				
 				if ( pUnit->GetNextCommand() == 0 )
 				{
-					const uint16_t wDir = pUnit->GetFrontDirection() == pUnit->GetDirection() ? wDirToPoint : wDirToPoint + 32768;
+					const uint16_t wDir = pUnit->GetFrontDirection() == pUnit->GetDirection() ? uint16_t( wDirToPoint ) : wDirToPoint + 32768;
 					theGroupLogic.UnitCommand( SAIUnitCmd( ACTION_COMMAND_GUARD, pUnit->GetCenterPlain(), wDir ), pUnit, false );
 				}
